refactor: 자식 컴포넌트 무한 리렌더링 문제 해결
